Refactor database connection to improve clarity
